Add submodules to sister repositories
Adds the dependency on GeoJsonCityBuilder as a submodule in the packages-folder (instead of using the "git-package" system), which allows for easier development of the GeoJsonCityBuilder project. Added the geojson-data repository also as a submodule. In a subsequent commit, I will add a script to automatically import updated geojson-data from the submodule folder into the Unity project assets folder.